Add Array.at polyfill
