+++ src/components/App.tsx
@@ -1,5 +1,6 @@
 import React, { useContext } from "react";
 import { flushSync } from "react-dom";
+
 import { RoughCanvas } from "roughjs/bin/canvas";
 import rough from "roughjs/bin/rough";
 import clsx from "clsx";
@@ -6221,30 +6222,6 @@ if (
       },
     },
   });
-
-  if (process.env.NODE_ENV === ENV.TEST) {
-    // Taken from https://github.com/tc39/proposal-relative-indexing-method#polyfill so that it works in tests
-    /* eslint-disable */
-    Object.defineProperty(Array.prototype, "at", {
-      value: function (n: number) {
-        // ToInteger() abstract op
-        n = Math.trunc(n) || 0;
-        // Allow negative indexing from the end
-        if (n < 0) {
-          n += this.length;
-        }
-        // OOB access is guaranteed to return undefined
-        if (n < 0 || n >= this.length) {
-          return undefined;
-        }
-        // Otherwise, this is just normal property access
-        return this[n];
-      },
-      writable: true,
-      enumerable: false,
-      configurable: true,
-    });
-  }
 }
 
 export default App;
