+++ src/packages/excalidraw/entry.js
@@ -1,5 +1,7 @@
 import "./publicPath";
+import polyfill from "../../polyfill";
 
 import "../../../public/fonts.css";
 
+polyfill();
 export * from "./index";
